// DELIVERABLES/Source Code/src/app/shared/models/contact/contact-filter.model.ts
export class ContactFilter {
    prospectSource: string = null;
    jobTitle: string = null;
    fromDate: number;
    toDate: number;
    month: string = null;
    customerId: string = null;
    assignedEmployeeId: string = null;
    bussiness: string = null;
    hasUpcomingBirthday: boolean;
    assignTo: string = null;
    assignToGroup: string = null;
}